moved the logic to useFeaturedPlays hook
